Refatoracao do codigo -> criacao de arquivo de rotas, de controller e refatorando arquivos ja existentes.
