+++ package.json
@@ -6,8 +6,8 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "dev": "node --watch --no-warnings server.js"
+    "start": "node ./src/main/server.js",
+    "dev": "node --watch --trace-warnings ./src/main/server.js"
   },
   "keywords": [],
   "author": "",
