--- server.js
@@ -1,62 +1,0 @@
-import { fastify } from "fastify";
-import {DatabaseMemory} from "./database-memory.js";
-import { DatabasePostgres } from "./database-postgres.js";
-
-
-const server = fastify();
-
-//const database = new DatabaseMemory();
-
-const database = new DatabasePostgres();
-
-server.post('/videos', async (req, res) => {
-
-    const {title, description, duration} = req.body;
-
-    database.create({ 
-        title,
-        description,
-        duration,
-    });
-
-    return res.status(201).send();
-});
-
-server.get('/videos', async (req, res) => {
-    const search = req.query.search;
-    const videos = database.list(search);
-
-    return videos;
-});
-server.get('/videos/:id', async (req, res) => {
-    const id = req.params.id;
-    const video = database.getById(id);
-
-    return video;
-});
-
-server.put('/videos/:id', async (req, res) => {
-    const videoId = req.params.id;
-
-    const {title, description, duration} = req.body;
-
-    database.update(videoId, {
-        title,
-        description,
-        duration,
-    });
-
-    return res.status(204).send();
-});
-
-server.delete('/videos/:id', async (req, res) => {
-    const videoId = req.params.id;
-
-    database.delete(videoId);
-
-    return res.status(204).send();
-});
-
-server.listen({
-    port: 3333
-});
